unified gfx base class
