+++ cocos/core/gfx/command-allocator.ts
@@ -2,7 +2,7 @@
  * @category gfx
  */
 
-import { GFXObject, GFXObjectType } from './define';
+import { GFXObject, GFXObjectType, GFXStatus } from './define';
 import { GFXDevice } from './device';
 
 // tslint:disable-next-line:no-empty-interface
@@ -22,7 +22,18 @@ export abstract class GFXCommandAllocator extends GFXObject {
         this._device = device;
     }
 
-    public abstract initialize (info: IGFXCommandAllocatorInfo): boolean;
+    public initialize (info: IGFXCommandAllocatorInfo) {
+        if (this._initialize(info)) { this._status = GFXStatus.SUCCESS; return true; }
+        else { this._status = GFXStatus.FAILED; return false; }
+    }
 
-    public abstract destroy (): void;
+    public destroy () {
+        if (this._status !== GFXStatus.SUCCESS) { return; }
+        this._destroy();
+        this._status = GFXStatus.UNREADY;
+    }
+
+    protected abstract _initialize (info: IGFXCommandAllocatorInfo): boolean;
+
+    protected abstract _destroy (): void;
 }
