+++ cocos/core/gfx/render-pass.ts
@@ -10,6 +10,7 @@ import {
     GFXPipelineBindPoint,
     GFXStoreOp,
     GFXTextureLayout,
+    GFXStatus,
 } from './define';
 import { GFXDevice } from './device';
 
@@ -75,7 +76,20 @@ export abstract class GFXRenderPass extends GFXObject {
         this._device = device;
     }
 
-    public abstract initialize (info: IGFXRenderPassInfo): boolean;
+    public initialize (info: IGFXRenderPassInfo) {
+        this._colorInfos = info.colorAttachments || [];
+        this._depthStencilInfo = info.depthStencilAttachment || null;
+        if (this._initialize(info)) { this._status = GFXStatus.SUCCESS; return true; }
+        else { this._status = GFXStatus.FAILED; return false; }
+    }
 
-    public abstract destroy (): void;
+    public destroy () {
+        if (this._status !== GFXStatus.SUCCESS) { return; }
+        this._destroy();
+        this._status = GFXStatus.UNREADY;
+    }
+
+    protected abstract _initialize (info: IGFXRenderPassInfo): boolean;
+
+    protected abstract _destroy (): void;
 }
