+++ cocos/core/gfx/webgl/webgl-binding-layout.ts
@@ -1,6 +1,5 @@
-import { GFXBindingLayout, GFXBindingUnit, IGFXBindingLayoutInfo } from '../binding-layout';
-import { GFXBindingType, GFXStatus } from '../define';
-import { GFXDevice } from '../device';
+import { GFXBindingLayout, IGFXBindingLayoutInfo } from '../binding-layout';
+import { GFXBindingType } from '../define';
 import { WebGLGFXBuffer } from './webgl-buffer';
 import { WebGLGPUBinding, WebGLGPUBindingLayout } from './webgl-gpu-objects';
 import { WebGLGFXSampler } from './webgl-sampler';
@@ -14,24 +13,7 @@ export class WebGLGFXBindingLayout extends GFXBindingLayout {
 
     private _gpuBindingLayout: WebGLGPUBindingLayout | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXBindingLayoutInfo): boolean {
-
-        this._bindingUnits = new Array<GFXBindingUnit>(info.bindings.length);
-        for (let i = 0; i < info.bindings.length; ++i) {
-            const binding = info.bindings[i];
-            this._bindingUnits[i] = {
-                binding: binding.binding,
-                type: binding.bindingType,
-                name: binding.name,
-                buffer: null,
-                texView: null,
-                sampler: null,
-            };
-        }
+    protected _initialize (info: IGFXBindingLayoutInfo): boolean {
 
         this._gpuBindingLayout = {
             gpuBindings: new Array<WebGLGPUBinding>(info.bindings.length),
@@ -49,14 +31,11 @@ export class WebGLGFXBindingLayout extends GFXBindingLayout {
             };
         }
 
-        this._status = GFXStatus.SUCCESS;
-
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         this._gpuBindingLayout = null;
-        this._status = GFXStatus.UNREADY;
     }
 
     public update () {
