+++ cocos/core/gfx/webgl/webgl-command-buffer.ts
@@ -77,27 +77,18 @@ export class WebGLGFXCommandBuffer extends GFXCommandBuffer {
     protected _curStencilCompareMask: IWebGLStencilCompareMask | null = null;
     protected _isStateInvalied: boolean = false;
 
-    public initialize (info: IGFXCommandBufferInfo): boolean {
+    protected _initialize (info: IGFXCommandBufferInfo): boolean {
 
-        if (!info.allocator) {
-            return false;
-        }
-
-        this._allocator = info.allocator;
         this._webGLAllocator = this._allocator as WebGLGFXCommandAllocator;
-        this._type = info.type;
-        this._status = GFXStatus.SUCCESS;
 
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         if (this._webGLAllocator) {
             this._webGLAllocator.clearCmds(this.cmdPackage);
-            this._allocator = null;
             this._webGLAllocator = null;
         }
-        this._status = GFXStatus.UNREADY;
     }
 
     public begin () {
@@ -333,9 +324,16 @@ export class WebGLGFXCommandBuffer extends GFXCommandBuffer {
                 this.bindStates();
             }
 
-            const cmd = ( this._allocator as WebGLGFXCommandAllocator).
+            const cmd = (this._allocator as WebGLGFXCommandAllocator).
                         drawCmdPool.alloc(WebGLCmdDraw);
-            (inputAssembler as WebGLGFXInputAssembler).extractCmdDraw(cmd);
+            // cmd.drawInfo = inputAssembler;
+            cmd.drawInfo.vertexCount = inputAssembler.vertexCount;
+            cmd.drawInfo.firstVertex = inputAssembler.firstVertex;
+            cmd.drawInfo.indexCount = inputAssembler.indexCount;
+            cmd.drawInfo.firstIndex = inputAssembler.firstIndex;
+            cmd.drawInfo.vertexOffset = inputAssembler.vertexOffset;
+            cmd.drawInfo.instanceCount = inputAssembler.instanceCount;
+            cmd.drawInfo.firstInstance = inputAssembler.firstInstance;
             this.cmdPackage.drawCmds.push(cmd);
 
             this.cmdPackage.cmds.push(WebGLCmd.DRAW);
