+++ cocos/core/gfx/webgl/webgl-framebuffer.ts
@@ -15,16 +15,7 @@ export class WebGLGFXFramebuffer extends GFXFramebuffer {
 
     private _gpuFramebuffer: WebGLGPUFramebuffer | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXFramebufferInfo): boolean {
-
-        this._renderPass = info.renderPass;
-        this._colorViews = info.colorViews || [];
-        this._depthStencilView = info.depthStencilView || null;
-        this._isOffscreen = info.isOffscreen !== undefined ? info.isOffscreen : true;
+    protected _initialize (info: IGFXFramebufferInfo): boolean {
 
         if (this._isOffscreen) {
 
@@ -59,16 +50,13 @@ export class WebGLGFXFramebuffer extends GFXFramebuffer {
             };
         }
 
-        this._status = GFXStatus.SUCCESS;
-
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         if (this._isOffscreen && this._gpuFramebuffer) {
             WebGLCmdFuncDestroyFramebuffer(this._device as WebGLGFXDevice, this._gpuFramebuffer);
         }
         this._gpuFramebuffer = null;
-        this._status = GFXStatus.UNREADY;
     }
 }
