+++ cocos/core/gfx/webgl/webgl-pipeline-layout.ts
@@ -1,7 +1,4 @@
-import { GFXStatus } from '../define';
-import { GFXDevice } from '../device';
 import { GFXPipelineLayout, IGFXPipelineLayoutInfo } from '../pipeline-layout';
-import { WebGLGFXDevice } from './webgl-device';
 import { WebGLGPUPipelineLayout } from './webgl-gpu-objects';
 
 export class WebGLGFXPipelineLayout extends GFXPipelineLayout {
@@ -12,20 +9,10 @@ export class WebGLGFXPipelineLayout extends GFXPipelineLayout {
 
     private _gpuPipelineLayout: WebGLGPUPipelineLayout | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXPipelineLayoutInfo): boolean {
-
-        this._layouts = info.layouts;
-        this._pushConstantsRanges = info.pushConstantsRanges || [];
-        this._status = GFXStatus.SUCCESS;
-
+    protected _initialize (info: IGFXPipelineLayoutInfo): boolean {
         return true;
     }
 
-    public destroy () {
-        this._status = GFXStatus.UNREADY;
+    protected _destroy () {
     }
 }
