+++ cocos/core/gfx/webgl/webgl-queue.ts
@@ -1,10 +1,5 @@
 import { GFXCommandBuffer } from '../command-buffer';
-import { GFXStatus } from '../define';
-import { GFXDevice } from '../device';
 import { GFXQueue, IGFXQueueInfo } from '../queue';
-import { WebGLGFXCommandBuffer } from './webgl-command-buffer';
-import { WebGLCmdFuncExecuteCmds } from './webgl-commands';
-import { WebGLGFXDevice } from './webgl-device';
 
 export class WebGLGFXQueue extends GFXQueue {
 
@@ -14,20 +9,12 @@ export class WebGLGFXQueue extends GFXQueue {
 
     private _isAsync: boolean = false;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXQueueInfo): boolean {
-
-        this._type = info.type;
-        this._status = GFXStatus.SUCCESS;
+    protected _initialize (info: IGFXQueueInfo): boolean {
 
         return true;
     }
 
-    public destroy () {
-        this._status = GFXStatus.UNREADY;
+    protected _destroy () {
     }
 
     public submit (cmdBuffs: GFXCommandBuffer[], fence?) {
