+++ cocos/core/gfx/webgl2/webgl2-buffer.ts
@@ -17,28 +17,8 @@ export class WebGL2GFXBuffer extends GFXBuffer {
     }
 
     private _gpuBuffer: WebGL2GPUBuffer | null = null;
-    private _indirectBuffer: IGFXIndirectBuffer | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXBufferInfo): boolean {
-
-        this._usage = info.usage;
-        this._memUsage = info.memUsage;
-        this._size = info.size;
-        this._stride = Math.max(info.stride || this._size, 1);
-        this._count = this._size / this._stride;
-        this._flags = (info.flags !== undefined ? info.flags : GFXBufferFlagBit.NONE);
-
-        if (this._usage & GFXBufferUsageBit.INDIRECT) {
-            this._indirectBuffer = { drawInfos: [] };
-        }
-
-        if (this._flags & GFXBufferFlagBit.BAKUP_BUFFER) {
-            this._bufferView = new Uint8Array(this._size);
-        }
+    protected _initialize (info: IGFXBufferInfo): boolean {
 
         this._gpuBuffer = {
             usage: info.usage,
@@ -58,13 +38,10 @@ export class WebGL2GFXBuffer extends GFXBuffer {
 
         WebGL2CmdFuncCreateBuffer(this._device as WebGL2GFXDevice, this._gpuBuffer);
 
-        this._device.memoryStatus.bufferSize += this._size;
-        this._status = GFXStatus.SUCCESS;
-
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         if (this._gpuBuffer) {
             WebGL2CmdFuncDestroyBuffer(this._device as WebGL2GFXDevice, this._gpuBuffer);
             this._device.memoryStatus.bufferSize -= this._size;
@@ -72,40 +49,24 @@ export class WebGL2GFXBuffer extends GFXBuffer {
         }
 
         this._bufferView = null;
-        this._status = GFXStatus.UNREADY;
     }
 
-    public resize (size: number) {
-        const oldSize = this._size;
-        this._size = size;
-        this._count = this._size / this._stride;
-
-        if (this._bufferView && oldSize !== size) {
-            const oldView = this._bufferView;
-            this._bufferView = new Uint8Array(this._size);
-            this._bufferView.set(oldView);
-            if (this._gpuBuffer) {
+    protected _resize (size: number) {
+        if (this._gpuBuffer) {
+            if (this._bufferView) {
                 this._gpuBuffer.buffer = this._bufferView;
             }
-        }
 
-        if (this._gpuBuffer) {
-            this._gpuBuffer.size = this._size;
-            if (this._size > 0) {
+            this._gpuBuffer.size = size;
+            if (size > 0) {
                 WebGL2CmdFuncResizeBuffer(this._device as WebGL2GFXDevice, this._gpuBuffer);
-                this._device.memoryStatus.bufferSize -= oldSize;
-                this._device.memoryStatus.bufferSize += this._size;
-                if (this._bufferView) {
-                    this._device.memoryStatus.bufferSize -= oldSize;
-                    this._device.memoryStatus.bufferSize += this._size;
-                }
             }
         }
     }
 
     public update (buffer: GFXBufferSource, offset?: number, size?: number) {
 
-        let buffSize;
+        let buffSize: number;
         if (size !== undefined ) {
             buffSize = size;
         } else if (this._usage & GFXBufferUsageBit.INDIRECT) {
