+++ cocos/core/gfx/webgl2/webgl2-command-allocator.ts
@@ -1,6 +1,5 @@
 import { CachedArray } from '../../memop/cached-array';
 import { GFXCommandAllocator, IGFXCommandAllocatorInfo } from '../command-allocator';
-import { GFXStatus } from '../define';
 import { GFXDevice } from '../device';
 import {
     WebGL2CmdBeginRenderPass,
@@ -14,7 +13,7 @@ import {
 
 export class WebGL2GFXCommandPool<T extends WebGL2CmdObject> {
 
-    private _frees: Array<T|null>;
+    private _frees: (T|null)[];
     private _freeIdx: number = 0;
     private _freeCmds: CachedArray<T>;
 
@@ -99,13 +98,11 @@ export class WebGL2GFXCommandAllocator extends GFXCommandAllocator {
         this.copyBufferToTextureCmdPool = new WebGL2GFXCommandPool(WebGL2CmdCopyBufferToTexture, 1);
     }
 
-    public initialize (info: IGFXCommandAllocatorInfo): boolean {
-        this._status = GFXStatus.SUCCESS;
+    protected _initialize (info: IGFXCommandAllocatorInfo): boolean {
         return true;
     }
 
-    public destroy () {
-        this._status = GFXStatus.UNREADY;
+    protected _destroy () {
     }
 
     public clearCmds (cmdPackage: WebGL2CmdPackage) {
