+++ cocos/core/gfx/webgl2/webgl2-command-buffer.ts
@@ -77,27 +77,18 @@ export class WebGL2GFXCommandBuffer extends GFXCommandBuffer {
     protected _curStencilCompareMask: IWebGL2StencilCompareMask | null = null;
     protected _isStateInvalied: boolean = false;
 
-    public initialize (info: IGFXCommandBufferInfo): boolean {
+    protected _initialize (info: IGFXCommandBufferInfo): boolean {
 
-        if (!info.allocator) {
-            return false;
-        }
-
-        this._allocator = info.allocator;
         this._webGLAllocator = this._allocator as WebGL2GFXCommandAllocator;
-        this._type = info.type;
-        this._status = GFXStatus.SUCCESS;
 
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         if (this._webGLAllocator) {
             this._webGLAllocator.clearCmds(this.cmdPackage);
-            this._allocator = null;
             this._webGLAllocator = null;
         }
-        this._status = GFXStatus.UNREADY;
     }
 
     public begin () {
@@ -332,9 +323,16 @@ export class WebGL2GFXCommandBuffer extends GFXCommandBuffer {
                 this.bindStates();
             }
 
-            const cmd = ( this._allocator as WebGL2GFXCommandAllocator).
+            const cmd = (this._allocator as WebGL2GFXCommandAllocator).
                         drawCmdPool.alloc(WebGL2CmdDraw);
-            (inputAssembler as WebGL2GFXInputAssembler).extractCmdDraw(cmd);
+            // cmd.drawInfo = inputAssembler;
+            cmd.drawInfo.vertexCount = inputAssembler.vertexCount;
+            cmd.drawInfo.firstVertex = inputAssembler.firstVertex;
+            cmd.drawInfo.indexCount = inputAssembler.indexCount;
+            cmd.drawInfo.firstIndex = inputAssembler.firstIndex;
+            cmd.drawInfo.vertexOffset = inputAssembler.vertexOffset;
+            cmd.drawInfo.instanceCount = inputAssembler.instanceCount;
+            cmd.drawInfo.firstInstance = inputAssembler.firstInstance;
             this.cmdPackage.drawCmds.push(cmd);
 
             this.cmdPackage.cmds.push(WebGL2Cmd.DRAW);
