+++ cocos/core/gfx/webgl2/webgl2-input-assembler.ts
@@ -1,8 +1,6 @@
-import { GFXStatus } from '../define';
-import { GFXDevice } from '../device';
 import { GFXInputAssembler, IGFXInputAssemblerInfo } from '../input-assembler';
 import { WebGL2GFXBuffer } from './webgl2-buffer';
-import { WebGL2CmdDraw, WebGL2CmdFuncCreateInputAssember, WebGL2CmdFuncDestroyInputAssembler } from './webgl2-commands';
+import { WebGL2CmdFuncCreateInputAssember, WebGL2CmdFuncDestroyInputAssembler } from './webgl2-commands';
 import { WebGL2GFXDevice } from './webgl2-device';
 import { IWebGL2GPUInputAssembler, WebGL2GPUBuffer } from './webgl2-gpu-objects';
 
@@ -14,29 +12,7 @@ export class WebGL2GFXInputAssembler extends GFXInputAssembler {
 
     private _gpuInputAssembler: IWebGL2GPUInputAssembler | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXInputAssemblerInfo): boolean {
-
-        if (info.vertexBuffers.length === 0) {
-            console.error('GFXInputAssemblerInfo.vertexBuffers is null.');
-            return false;
-        }
-
-        this._attributes = info.attributes;
-        this._vertexBuffers = info.vertexBuffers;
-
-        if (info.indexBuffer !== undefined) {
-            this._indexBuffer = info.indexBuffer;
-            this._indexCount = this._indexBuffer.size / this._indexBuffer.stride;
-        } else {
-            const vertBuff = this._vertexBuffers[0];
-            this._vertexCount = vertBuff.size / vertBuff.stride;
-        }
-
-        this._indirectBuffer = info.indirectBuffer || null;
+    public _initialize (info: IGFXInputAssemblerInfo): boolean {
 
         const gpuVertexBuffers: WebGL2GPUBuffer[] = new Array<WebGL2GPUBuffer>(info.vertexBuffers.length);
         for (let i = 0; i < info.vertexBuffers.length; ++i) {
@@ -80,27 +56,14 @@ export class WebGL2GFXInputAssembler extends GFXInputAssembler {
 
         WebGL2CmdFuncCreateInputAssember(this._device as WebGL2GFXDevice, this._gpuInputAssembler);
 
-        this._status = GFXStatus.SUCCESS;
-
         return true;
     }
 
-    public destroy () {
+    public _destroy () {
         const webgl2Dev = this._device as WebGL2GFXDevice;
         if (this._gpuInputAssembler && webgl2Dev.useVAO) {
             WebGL2CmdFuncDestroyInputAssembler(webgl2Dev, this._gpuInputAssembler);
         }
         this._gpuInputAssembler = null;
-        this._status = GFXStatus.UNREADY;
-    }
-
-    public extractCmdDraw (cmd: WebGL2CmdDraw) {
-        cmd.drawInfo.vertexCount = this._vertexCount;
-        cmd.drawInfo.firstVertex = this._firstVertex;
-        cmd.drawInfo.indexCount = this._indexCount;
-        cmd.drawInfo.firstIndex = this._firstIndex;
-        cmd.drawInfo.vertexOffset = this._vertexOffset;
-        cmd.drawInfo.instanceCount = this._instanceCount;
-        cmd.drawInfo.firstInstance = this._firstInstance;
     }
 }
