+++ cocos/core/gfx/webgl2/webgl2-pipeline-layout.ts
@@ -1,5 +1,3 @@
-import { GFXStatus } from '../define';
-import { GFXDevice } from '../device';
 import { GFXPipelineLayout, IGFXPipelineLayoutInfo } from '../pipeline-layout';
 import { WebGL2GPUPipelineLayout } from './webgl2-gpu-objects';
 
@@ -11,20 +9,10 @@ export class WebGL2GFXPipelineLayout extends GFXPipelineLayout {
 
     private _gpuPipelineLayout: WebGL2GPUPipelineLayout | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXPipelineLayoutInfo): boolean {
-
-        this._layouts = info.layouts;
-        this._pushConstantsRanges = info.pushConstantsRanges || [];
-        this._status = GFXStatus.SUCCESS;
-
+    protected _initialize (info: IGFXPipelineLayoutInfo): boolean {
         return true;
     }
 
-    public destroy () {
-        this._status = GFXStatus.UNREADY;
+    protected _destroy () {
     }
 }
