+++ cocos/core/gfx/webgl2/webgl2-pipeline-state.ts
@@ -1,5 +1,3 @@
-import { GFXStatus } from '../define';
-import { GFXDevice } from '../device';
 import { GFXPipelineState, IGFXPipelineStateInfo } from '../pipeline-state';
 import { WebGL2GPUPipelineState } from './webgl2-gpu-objects';
 import { WebGL2GFXPipelineLayout } from './webgl2-pipeline-layout';
@@ -31,23 +29,7 @@ export class WebGL2GFXPipelineState extends GFXPipelineState {
 
     private _gpuPipelineState: WebGL2GPUPipelineState | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXPipelineStateInfo): boolean {
-
-        this._primitive = info.primitive;
-        this._shader = info.shader;
-        this._is = info.inputState;
-        this._rs = info.rasterizerState;
-        this._dss = info.depthStencilState;
-        this._bs = info.blendState;
-        this._dynamicStates = info.dynamicStates || [];
-        this._hash = info.hash;
-
-        this._layout = info.layout;
-        this._renderPass = info.renderPass;
+    protected _initialize (info: IGFXPipelineStateInfo): boolean {
 
         this._gpuPipelineState = {
             glPrimitive: WebGLPrimitives[info.primitive],
@@ -60,13 +42,11 @@ export class WebGL2GFXPipelineState extends GFXPipelineState {
             gpuRenderPass: (info.renderPass as WebGL2GFXRenderPass).gpuRenderPass,
         };
 
-        this._status = GFXStatus.SUCCESS;
 
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         this._gpuPipelineState = null;
-        this._status = GFXStatus.UNREADY;
     }
 }
