+++ cocos/core/gfx/webgl2/webgl2-queue.ts
@@ -1,10 +1,6 @@
 import { GFXCommandBuffer } from '../command-buffer';
-import { GFXStatus } from '../define';
-import { GFXDevice } from '../device';
 import { GFXQueue, IGFXQueueInfo } from '../queue';
 import { WebGL2GFXCommandBuffer } from './webgl2-command-buffer';
-import { WebGL2CmdFuncExecuteCmds } from './webgl2-commands';
-import { WebGL2GFXDevice } from './webgl2-device';
 
 export class WebGL2GFXQueue extends GFXQueue {
 
@@ -14,20 +10,12 @@ export class WebGL2GFXQueue extends GFXQueue {
 
     private _isAsync: boolean = false;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXQueueInfo): boolean {
-
-        this._type = info.type;
-        this._status = GFXStatus.SUCCESS;
+    protected _initialize (info: IGFXQueueInfo): boolean {
 
         return true;
     }
 
-    public destroy () {
-        this._status = GFXStatus.UNREADY;
+    protected _destroy () {
     }
 
     public submit (cmdBuffs: GFXCommandBuffer[], fence?) {
