+++ cocos/core/gfx/webgl2/webgl2-sampler.ts
@@ -1,6 +1,4 @@
-import { GFXStatus } from '../define';
-import { GFXDevice } from '../device';
-import { GFXSampler, GFXSamplerState, IGFXSamplerInfo } from '../sampler';
+import { GFXSampler, IGFXSamplerInfo } from '../sampler';
 import { WebGL2CmdFuncCreateSampler, WebGL2CmdFuncDestroySampler } from './webgl2-commands';
 import { WebGL2GFXDevice } from './webgl2-device';
 import { WebGL2GPUSampler } from './webgl2-gpu-objects';
@@ -13,65 +11,7 @@ export class WebGL2GFXSampler extends GFXSampler {
 
     private _gpuSampler: WebGL2GPUSampler | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-
-        this._state = new GFXSamplerState();
-    }
-
-    public initialize (info: IGFXSamplerInfo): boolean {
-
-        if (info.name !== undefined) {
-            this._state.name = info.name;
-        }
-
-        if (info.minFilter !== undefined) {
-            this._state.minFilter = info.minFilter;
-        }
-
-        if (info.magFilter !== undefined) {
-            this._state.magFilter = info.magFilter;
-        }
-
-        if (info.mipFilter !== undefined) {
-            this._state.mipFilter = info.mipFilter;
-        }
-
-        if (info.addressU !== undefined) {
-            this._state.addressU = info.addressU;
-        }
-
-        if (info.addressV !== undefined) {
-            this._state.addressV = info.addressV;
-        }
-
-        if (info.addressW !== undefined) {
-            this._state.addressW = info.addressW;
-        }
-
-        if (info.maxAnisotropy !== undefined) {
-            this._state.maxAnisotropy = info.maxAnisotropy;
-        }
-
-        if (info.cmpFunc !== undefined) {
-            this._state.cmpFunc = info.cmpFunc;
-        }
-
-        if (info.borderColor !== undefined) {
-            this._state.borderColor = info.borderColor;
-        }
-
-        if (info.minLOD !== undefined) {
-            this._state.minLOD = info.minLOD;
-        }
-
-        if (info.maxLOD !== undefined) {
-            this._state.maxLOD = info.maxLOD;
-        }
-
-        if (info.mipLODBias !== undefined) {
-            this._state.mipLODBias = info.mipLODBias;
-        }
+    protected _initialize (info: IGFXSamplerInfo): boolean {
 
         this._gpuSampler = {
             glSampler: null,
@@ -93,16 +33,13 @@ export class WebGL2GFXSampler extends GFXSampler {
 
         WebGL2CmdFuncCreateSampler(this._device as WebGL2GFXDevice, this._gpuSampler);
 
-        this._status = GFXStatus.SUCCESS;
-
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         if (this._gpuSampler) {
             WebGL2CmdFuncDestroySampler(this._device as WebGL2GFXDevice, this._gpuSampler);
             this._gpuSampler = null;
         }
-        this._status = GFXStatus.UNREADY;
     }
 }
