+++ cocos/core/gfx/webgl2/webgl2-texture.ts
@@ -1,13 +1,8 @@
-import { GFXFormatSurfaceSize, GFXStatus, GFXTextureFlagBit, GFXTextureType, GFXTextureViewType } from '../define';
-import { GFXDevice } from '../device';
+import { GFXTextureFlagBit, GFXTextureType, GFXTextureViewType } from '../define';
 import { GFXTexture, IGFXTextureInfo } from '../texture';
 import { WebGL2CmdFuncCreateTexture, WebGL2CmdFuncDestroyTexture, WebGL2CmdFuncResizeTexture } from './webgl2-commands';
 import { WebGL2GFXDevice } from './webgl2-device';
 import { WebGL2GPUTexture } from './webgl2-gpu-objects';
-
-function IsPowerOf2 (x: number): boolean{
-    return x > 0 && (x & (x - 1)) === 0;
-}
 
 export class WebGL2GFXTexture extends GFXTexture {
 
@@ -17,46 +12,7 @@ export class WebGL2GFXTexture extends GFXTexture {
 
     private _gpuTexture: WebGL2GPUTexture | null = null;
 
-    constructor (device: GFXDevice) {
-        super(device);
-    }
-
-    public initialize (info: IGFXTextureInfo): boolean {
-
-        this._type = info.type;
-        this._usage = info.usage;
-        this._format = info.format;
-        this._width = info.width;
-        this._height = info.height;
-
-        if (info.depth !== undefined) {
-            this._depth = info.depth;
-        }
-
-        if (info.arrayLayer !== undefined) {
-            this._arrayLayer = info.arrayLayer;
-        }
-
-        if (info.mipLevel !== undefined) {
-            this._mipLevel = info.mipLevel;
-        }
-
-        if (info.samples !== undefined) {
-            this._samples = info.samples;
-        }
-
-        if (info.flags !== undefined) {
-            this._flags = info.flags;
-        }
-
-        this._isPowerOf2 = IsPowerOf2(this._width) && IsPowerOf2(this._height);
-
-        this._size = GFXFormatSurfaceSize(this._format, this.width, this.height,
-            this.depth, this.mipLevel) * this._arrayLayer;
-
-        if (this._flags & GFXTextureFlagBit.BAKUP_BUFFER) {
-            this._buffer = new ArrayBuffer(this._size);
-        }
+    protected _initialize (info: IGFXTextureInfo): boolean {
 
         let viewType: GFXTextureViewType;
         switch (info.type) {
@@ -129,37 +85,22 @@ export class WebGL2GFXTexture extends GFXTexture {
 
         WebGL2CmdFuncCreateTexture(this._device as WebGL2GFXDevice, this._gpuTexture);
 
-        this._device.memoryStatus.textureSize += this._size;
-        this._status = GFXStatus.SUCCESS;
-
         return true;
     }
 
-    public destroy () {
+    protected _destroy () {
         if (this._gpuTexture) {
             WebGL2CmdFuncDestroyTexture(this._device as WebGL2GFXDevice, this._gpuTexture);
-            this._device.memoryStatus.textureSize -= this._size;
             this._gpuTexture = null;
         }
-        this._status = GFXStatus.UNREADY;
-        this._buffer = null;
     }
 
-    public resize (width: number, height: number) {
-        const oldSize = this._size;
-        this._width = width;
-        this._height = height;
-        this._size = GFXFormatSurfaceSize(this._format, this.width, this.height,
-            this.depth, this.mipLevel) * this._arrayLayer;
-
+    protected _resize (width: number, height: number) {
         if (this._gpuTexture) {
-            this._gpuTexture.width = this._width;
-            this._gpuTexture.height = this._height;
+            this._gpuTexture.width = width;
+            this._gpuTexture.height = height;
             this._gpuTexture.size = this._size;
             WebGL2CmdFuncResizeTexture(this._device as WebGL2GFXDevice, this._gpuTexture);
-            this._device.memoryStatus.textureSize -= oldSize;
-            this._device.memoryStatus.textureSize += this._size;
         }
-        this._status = GFXStatus.UNREADY;
     }
 }
